requirements.txt
Django==5.1.6
redis==5.2.1
dj_database_url==2.3.0
whitenoise==6.9.0
celery==5.4.0
django-environ
gunicorn==23.0.0
psycopg2==2.9.10
djangorestframework==3.15.2
django-admin-interface==0.29.4
 django-admin-autocomplete-filter==0.7.1
 django-filter==25.1
 djangorestframework-simplejwt==5.4.0